aligned with v4.0.0 of spec
- removed requests from reversals so they are BasicReversals
- required adding a flag to methods to ID payment vs refund reversals
for backend
- moved balance and request amounts from old locations to Amounts model
