+++ pom.xml
@@ -4,7 +4,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>io.electrum</groupId>
   <artifactId>billpay-test-server</artifactId>
-  <version>2.0.0</version>
+  <version>2.0.1</version>
   <packaging>jar</packaging>
   <licenses>
     <license>
@@ -34,6 +34,7 @@
     <jetty-version>9.2.18.v20160721</jetty-version>
     <jersey-version>2.22.1</jersey-version>
     <joda-time-version>2.9.4</joda-time-version>
+    <billpay-version>4.0.0</billpay-version>
   </properties>
 
   <dependencies>
@@ -42,7 +43,7 @@
     <dependency>
       <groupId>io.electrum</groupId>
       <artifactId>billpay-service-interface</artifactId>
-      <version>3.0.0</version>
+      <version>${billpay-version}</version>
     </dependency>
     
     <!-- CSV -->
