+++ src/main/java/io/electrum/billpaytestserver/BillpayTestServerRunner.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.electrum.billpaytestserver;
+package io.electrum.billpaytestserver;
 
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.HttpConfiguration;
