+++ src/main/java/io/electrum/billpaytestserver/engine/ErrorDetailFactory.java
@@ -1,14 +1,13 @@
-package com.electrum.billpaytestserver.engine;
+package io.electrum.billpaytestserver.engine;
 
+import io.electrum.billpay.model.ErrorDetail;
+import io.electrum.billpaytestserver.account.BillPayAccount;
+import io.electrum.billpaytestserver.validation.ValidationResult;
 import io.electrum.vas.model.BasicAdvice;
-import io.electrum.vas.model.ErrorDetail;
 
 import java.util.UUID;
 
 import javax.ws.rs.core.Response;
-
-import com.electrum.billpaytestserver.account.BillPayAccount;
-import com.electrum.billpaytestserver.validation.ValidationResult;
 
 /**
  *
