+++ src/main/java/io/electrum/billpaytestserver/handler/RefundResourceHandler.java
@@ -1,11 +1,4 @@
-package com.electrum.billpaytestserver.handler;
-
-import io.electrum.billpay.api.IPaymentsResource;
-import io.electrum.billpay.model.PaymentRequest;
-import io.electrum.billpay.model.PaymentResponse;
-import io.electrum.billpay.model.PaymentReversal;
-import io.electrum.vas.model.LedgerAmount;
-import io.electrum.vas.model.TenderAdvice;
+package io.electrum.billpaytestserver.handler;
 
 import java.util.UUID;
 
@@ -20,42 +13,52 @@ import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.electrum.billpaytestserver.Utils;
-import com.electrum.billpaytestserver.account.BillPayAccount;
-import com.electrum.billpaytestserver.engine.ErrorDetailFactory;
-import com.electrum.billpaytestserver.engine.MockBillPayBackend;
 import com.fasterxml.jackson.core.JsonProcessingException;
+
+import io.electrum.billpay.api.IRefundsResource;
+import io.electrum.billpay.model.PaymentResponse;
+import io.electrum.billpay.model.RefundRequest;
+import io.electrum.billpay.model.RefundResponse;
+import io.electrum.billpaytestserver.Utils;
+import io.electrum.billpaytestserver.account.BillPayAccount;
+import io.electrum.billpaytestserver.engine.ErrorDetailFactory;
+import io.electrum.billpaytestserver.engine.MockBillPayBackend;
+import io.electrum.vas.model.Amounts;
+import io.electrum.vas.model.BasicAdvice;
+import io.electrum.vas.model.BasicReversal;
+import io.electrum.vas.model.LedgerAmount;
 
 /**
  *
  */
-public class PaymentResourceHandler extends BaseRequestHandler<PaymentRequest, PaymentResponse>
-      implements IPaymentsResource {
-   private static final Logger log = LoggerFactory.getLogger(PaymentResourceHandler.class);
+public class RefundResourceHandler extends BaseRequestHandler<RefundRequest, RefundResponse>
+      implements IRefundsResource {
+   private static final Logger log = LoggerFactory.getLogger(RefundResourceHandler.class);
 
    @Override
-   public void confirmPayment(
+   public void confirmRefund(
          UUID adviceId,
-         UUID paymentId,
-         TenderAdvice tenderAdvice,
+         UUID refundId,
+         BasicAdvice basicAdvice,
          SecurityContext securityContext,
          AsyncResponse asyncResponse,
          Request request,
          HttpServletRequest httpServletRequest,
          HttpHeaders httpHeaders,
          UriInfo uriInfo) {
-      log.info("Handling payment confirm");
+      log.info("Handling refund confirm");
       try {
          handleConfirm(
                adviceId,
-               paymentId,
-               tenderAdvice,
+               refundId,
+               basicAdvice,
                securityContext,
                asyncResponse,
                request,
                httpServletRequest,
                httpHeaders,
-               uriInfo);
+               uriInfo,
+               false);
       } catch (Exception e) {
          log.error("Error handling message", e);
          asyncResponse.resume(ErrorDetailFactory.getServerErrorErrorDetail(e));
@@ -63,20 +66,20 @@ public class PaymentResourceHandler extends BaseRequestHandler<PaymentRequest, P
    }
 
    @Override
-   public void createPayment(
+   public void createRefund(
          UUID uuid,
-         PaymentRequest paymentRequest,
+         RefundRequest refundRequest,
          SecurityContext securityContext,
          AsyncResponse asyncResponse,
          Request request,
          HttpServletRequest httpServletRequest,
          HttpHeaders httpHeaders,
          UriInfo uriInfo) {
-      log.info("Handling payment request");
+      log.info("Handling refund request");
       try {
          handleMessage(
                uuid,
-               paymentRequest,
+               refundRequest,
                securityContext,
                asyncResponse,
                request,
@@ -90,52 +93,68 @@ public class PaymentResourceHandler extends BaseRequestHandler<PaymentRequest, P
    }
 
    @Override
-   public void reversePayment(
+   public void reverseRefund(
          UUID adviceId,
-         UUID paymentId,
-         PaymentReversal paymentReversal,
+         UUID refundId,
+         BasicReversal refundReversal,
          SecurityContext securityContext,
          AsyncResponse asyncResponse,
          Request request,
          HttpServletRequest httpServletRequest,
          HttpHeaders httpHeaders,
          UriInfo uriInfo) {
-      log.info("Handling payment reversal");
+      log.info("Handling refund reversal");
       try {
          handleReversal(
                adviceId,
-               paymentId,
-               paymentReversal,
+               refundId,
+               refundReversal,
                securityContext,
                asyncResponse,
                request,
                httpServletRequest,
                httpHeaders,
-               uriInfo);
+               uriInfo,
+               false);
       } catch (Exception e) {
          log.error("Error handling message", e);
          asyncResponse.resume(ErrorDetailFactory.getServerErrorErrorDetail(e));
       }
+
    }
 
-   protected void doConfirm(PaymentRequest request) {
-      BillPayAccount account = MockBillPayBackend.getAccount(request.getAccountRef());
+   protected void doConfirm(RefundRequest request) {
+
+      PaymentResponse origPaymentResponse = MockBillPayBackend.getPaymentResponse(request.getIssuerReference());
+
+      BillPayAccount account = MockBillPayBackend.getAccount(origPaymentResponse.getAccount().getAccountRef());
 
       LedgerAmount ledgerAmount = account.getBalance();
 
       long amount = ledgerAmount.getAmount();
 
-      amount -= request.getRequestAmount().getAmount();
+      amount += origPaymentResponse.getAmounts().getApprovedAmount().getAmount();
 
       ledgerAmount.setAmount(amount);
    }
 
-   protected void doReversal(PaymentRequest request) {
+   protected void doReversal(RefundRequest request) {
+      PaymentResponse origPaymentResponse = MockBillPayBackend.getPaymentResponse(request.getIssuerReference());
+
+      BillPayAccount account = MockBillPayBackend.getAccount(origPaymentResponse.getAccount().getAccountRef());
+
+      LedgerAmount ledgerAmount = account.getBalance();
+
+      long amount = ledgerAmount.getAmount();
+
+      amount -= origPaymentResponse.getAmounts().getApprovedAmount().getAmount();
+
+      ledgerAmount.setAmount(amount);
    }
 
-   protected PaymentResponse getResponse(PaymentRequest request, BillPayAccount account) {
+   protected RefundResponse getResponse(RefundRequest request, BillPayAccount account) {
       log.info("Constructing response");
-      PaymentResponse response = new PaymentResponse();
+      RefundResponse response = new RefundResponse();
 
       response.setId(request.getId());
       response.setTime(new DateTime());
@@ -144,9 +163,8 @@ public class PaymentResourceHandler extends BaseRequestHandler<PaymentRequest, P
       response.setSettlementEntity(getSettlementEntity());
       response.setReceiver(getReceiver());
       response.setAccount(getAccount(account));
+      response.setAmounts(new Amounts().balanceAmount(account.getBalance()));
       response.setCustomer(account.getCustomer());
-      response.setSlipData(getSlipData());
-      response.setResponseAmount(request.getRequestAmount());
       response.setThirdPartyIdentifiers(getThirdPartyIdentifiers(request.getThirdPartyIdentifiers()));
 
       try {
@@ -154,8 +172,6 @@ public class PaymentResourceHandler extends BaseRequestHandler<PaymentRequest, P
       } catch (JsonProcessingException e) {
          log.error("Could not print response");
       }
-
-      MockBillPayBackend.add(response);
 
       return response;
    }
