+++ src/main/java/io/electrum/billpaytestserver/validation/BillpayMessageValidator.java
@@ -1,16 +1,13 @@
-package com.electrum.billpaytestserver.validation;
+package io.electrum.billpaytestserver.validation;
 
 import io.electrum.billpay.model.AccountLookupRequest;
 import io.electrum.billpay.model.PaymentRequest;
-import io.electrum.billpay.model.PaymentReversal;
-import io.electrum.billpay.model.RefundReversal;
 import io.electrum.vas.model.BasicAdvice;
+import io.electrum.vas.model.BasicReversal;
 import io.electrum.vas.model.Institution;
-import io.electrum.vas.model.LedgerAmount;
 import io.electrum.vas.model.Merchant;
 import io.electrum.vas.model.MerchantName;
 import io.electrum.vas.model.Originator;
-import io.electrum.vas.model.Tender;
 import io.electrum.vas.model.TenderAdvice;
 import io.electrum.vas.model.ThirdPartyIdentifier;
 import io.electrum.vas.model.Transaction;
@@ -36,14 +33,8 @@ public class BillpayMessageValidator {
 
       validate(request, result);
 
-      if (request instanceof AccountLookupRequest) {
+      if (request instanceof AccountLookupRequest || request instanceof PaymentRequest) {
          validateValue(request, "message", "accountRef", result);
-      }
-
-      if (request instanceof PaymentRequest) {
-         validateValue(request, "message", "accountRef", result);
-         validateValue(request, "message", "requestAmount", result);
-         validate(((PaymentRequest) request).getRequestAmount(), result);
       }
 
       return result;
@@ -58,10 +49,8 @@ public class BillpayMessageValidator {
 
       if (advice instanceof TenderAdvice) {
          validate((TenderAdvice) advice, result);
-      } else if (advice instanceof PaymentReversal) {
-         validate((PaymentReversal) advice, result);
-      } else if (advice instanceof RefundReversal) {
-         validate((RefundReversal) advice, result);
+      } else if (advice instanceof BasicReversal) {
+         validate((BasicReversal) advice, result);
       }
 
       validate(advice, result);
@@ -71,25 +60,10 @@ public class BillpayMessageValidator {
 
    private static void validate(TenderAdvice advice, ValidationResult result) {
       validateValue(advice, "message", "tenders", result);
-      for (Tender tender : advice.getTenders()) {
-         validate(tender, result);
-      }
-
    }
 
-   public static void validate(PaymentReversal reversal, ValidationResult result) {
+   public static void validate(BasicReversal reversal, ValidationResult result) {
       validateValue(reversal, "message", "reversalReason", result);
-      validateValue(reversal, "message", "paymentRequest", result);
-      validateValue(reversal.getPaymentRequest(), "paymentRequest", "accountRef", result);
-      validate(reversal.getPaymentRequest(), result);
-   }
-
-   public static void validate(RefundReversal reversal, ValidationResult result) {
-      validateValue(reversal, "message", "reversalReason", result);
-      validateValue(reversal, "message", "refundRequest", result);
-      validateValue(reversal.getRefundRequest(), "refundRequest", "issuerReference", result);
-      validateValue(reversal.getRefundRequest(), "refundRequest", "refundReason", result);
-      validate(reversal.getRefundRequest(), result);
    }
 
    private static void validate(BasicAdvice advice, ValidationResult result) {
@@ -155,31 +129,6 @@ public class BillpayMessageValidator {
 
       validateValue(thirdPartyIdentifier, "thirdPartyIdentifier", "institutionId", result);
       validateValue(thirdPartyIdentifier, "thirdPartyIdentifier", "transactionIdentifier", result);
-
-   }
-
-   private static void validate(Tender tender, ValidationResult result) {
-      if (isEmpty(tender)) {
-         return;
-      }
-
-      validateValue(tender, "tender", "accountType", result);
-      validateValue(tender, "tender", "amount", result);
-      validate(tender.getAmount(), result);
-      validateValue(tender, "tender", "cardNumber", result);
-      validateValue(tender, "tender", "reference", result);
-      validateValue(tender, "tender", "tenderType", result);
-
-   }
-
-   private static void validate(LedgerAmount amount, ValidationResult result) {
-      if (isEmpty(amount)) {
-         return;
-      }
-
-      validateValue(amount, "ledgerAmount", "amount", result);
-      validateValue(amount, "ledgerAmount", "currency", result);
-      validateValue(amount, "ledgerAmount", "ledgerIndicator", result);
 
    }
