+++ src/main/java/io/electrum/billpaytestserver/ws/AccountLookupResource.java
@@ -1,11 +1,10 @@
-package com.electrum.billpaytestserver.ws;
+package io.electrum.billpaytestserver.ws;
 
 import io.electrum.billpay.api.AccountLookupsResource;
 import io.electrum.billpay.api.IAccountLookupsResource;
+import io.electrum.billpaytestserver.handler.AccountLookResourceHandler;
 
 import javax.ws.rs.Path;
-
-import com.electrum.billpaytestserver.handler.AccountLookResourceHandler;
 
 @Path("/accountLookups/{requestId}")
 public class AccountLookupResource extends AccountLookupsResource {
