+++ test/postman/heroku.postman_environment.json
@@ -1,6 +1,6 @@
 {
-	"id": "f6e93569-8931-a899-f732-2374aba978eb",
-	"name": "BillPayInterface Test Environment",
+	"id": "c697f0bd-4670-7475-a636-baef9d2ce7e9",
+	"name": "heroku",
 	"values": [
 		{
 			"key": "url",
@@ -137,7 +137,7 @@
 		{
 			"key": "currency",
 			"type": "text",
-			"value": "710",
+			"value": "ZAR",
 			"enabled": true
 		},
 		{
