+++ test/postman/localhost.postman_environment.json
@@ -1,6 +1,6 @@
 {
-	"id": "f6e93569-8931-a899-f732-2374aba978eb",
-	"name": "BillPayInterface Test Environment",
+	"id": "3936166f-21ec-555b-a306-b18a97440cca",
+	"name": "localhost",
 	"values": [
 		{
 			"key": "url",
@@ -17,19 +17,19 @@
 		{
 			"key": "rrn",
 			"type": "text",
-			"value": "7e5cb7ba",
+			"value": "fd792311",
 			"enabled": true
 		},
 		{
 			"key": "reqDateTime",
 			"type": "text",
-			"value": "2016-08-22T00:05:12.340Z",
+			"value": "2016-09-23T06:31:41.494Z",
 			"enabled": true
 		},
 		{
 			"key": "requestId",
 			"type": "text",
-			"value": "879da094-f304-3b96-3590-4c670ea27766",
+			"value": "d106670f-0601-ee5f-5e77-eba89899e17d",
 			"enabled": true
 		},
 		{
@@ -137,13 +137,13 @@
 		{
 			"key": "currency",
 			"type": "text",
-			"value": "710",
+			"value": "ZAR",
 			"enabled": true
 		},
 		{
 			"key": "accountAmount",
 			"type": "text",
-			"value": "32643",
+			"value": "-16320",
 			"enabled": true
 		},
 		{
@@ -245,13 +245,13 @@
 		{
 			"key": "senderTranRef",
 			"type": "text",
-			"value": "ed6eef54",
+			"value": "705eee5b",
 			"enabled": true
 		},
 		{
 			"key": "clientTranRef",
 			"type": "text",
-			"value": "fac1b71a",
+			"value": "b1064e0b",
 			"enabled": true
 		},
 		{
@@ -263,7 +263,7 @@
 		{
 			"key": "thirdPartyTranRef1",
 			"type": "text",
-			"value": "ed6eef54",
+			"value": "705eee5b",
 			"enabled": true
 		},
 		{
@@ -275,7 +275,7 @@
 		{
 			"key": "thirdPartyTranRef2",
 			"type": "text",
-			"value": "fac1b71a",
+			"value": "b1064e0b",
 			"enabled": true
 		},
 		{
@@ -303,7 +303,7 @@
 			"enabled": true
 		}
 	],
-	"timestamp": 1471824318888,
+	"timestamp": 1474612301612,
 	"synced": false,
 	"syncedFilename": "",
 	"team": null,
